+++ CameraPlus/Behaviours/ScreenCameraBehaviour.cs
@@ -1,7 +1,10 @@
 ﻿using CameraPlus.Utilities;
 using System.Collections.Generic;
 using System.Linq;
+using CameraPlus.Configuration;
+using Klak.Spout;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 namespace CameraPlus.Behaviours
 {
@@ -15,7 +18,8 @@ namespace CameraPlus.Behaviours
         private List<CameraPlusBehaviour> _cameras;
         private RenderTexture _renderTexture;
         private RenderTexture _backgroundTexture;
-
+        private RenderTexture _spoutTexture;
+        
         private Material _dofMaterial = null;
         private Material _wipeMaterial = null;
         private Material _outlineMaterial = null;
@@ -23,6 +27,10 @@ namespace CameraPlus.Behaviours
         private Material _dotMateial = null;
 
         private Rect _screenRect;
+
+        public SpoutReceiver ScreenSpoutReceiver;
+        private Vector2 _spoutScreenOffset;
+        private Material _spoutMaterial = new Material(Plugin.cameraController.Shaders["BeatSaber/Unlit/Transparent"]);
         public void SetLayer(int layer)
         {
             _cam.depth = layer;
@@ -81,6 +89,34 @@ namespace CameraPlus.Behaviours
                 anisoLevel = 1,
                 useDynamicScale = false
             };
+            
+            _spoutScreenOffset = Vector2.zero;
+            if(PluginConfig.Instance.AvatarSpoutAutoConnect)
+                CreateSpoutScreen(PluginConfig.Instance.AvatarSpoutName);
+        }
+
+        public void CreateSpoutScreen(string spoutName)
+        {
+            if (!ScreenSpoutReceiver)
+            {
+                _spoutMaterial = new Material(Plugin.cameraController.Shaders["BeatSaber/Unlit/Transparent"]);
+                var spoutResources = SpoutResources.CreateInstance<SpoutResources>();
+                spoutResources.blitShader = Plugin.cameraController.Shaders["Hidden/Klak/Spout/Blit"];
+                
+                ScreenSpoutReceiver = this.gameObject.AddComponent<SpoutReceiver>();
+                ScreenSpoutReceiver.SetResources(spoutResources);
+                ScreenSpoutReceiver.sourceName = spoutName;
+                _spoutTexture = new RenderTexture(Screen.width, Screen.height, 0, RenderTextureFormat.ARGB32);
+                ScreenSpoutReceiver.targetTexture = _spoutTexture;
+                ScreenSpoutReceiver.targetMaterialProperty = "_MainTex";
+            }
+        }
+
+        public void DisposeSpoutScreen()
+        {
+            if(!ScreenSpoutReceiver) return;
+            Destroy(ScreenSpoutReceiver);
+            ScreenSpoutReceiver = null;
         }
 
         private void OnRenderImage(RenderTexture src, RenderTexture dest)
@@ -104,6 +140,20 @@ namespace CameraPlus.Behaviours
                         PostEffect.Wipe(c, c._camRenderTexture, dest, _wipeMaterial);
                     else
                         Graphics.Blit(c._camRenderTexture, dest);
+                }
+            }
+
+            if (ScreenSpoutReceiver)
+            {
+                if ((SceneManager.GetActiveScene().name == "MainMenu" && PluginConfig.Instance.AvatarSpoutInMenu) ||
+                    (SceneManager.GetActiveScene().name == "GameCore" && PluginConfig.Instance.AvatarSpoutInGame))
+                {
+                    _cam.pixelRect = _screenRect;
+
+                    _spoutMaterial.SetFloat("_Threshold", PluginConfig.Instance.SpoutCameraAlpha);
+                    _spoutScreenOffset.x = PluginConfig.Instance.AvatarSpoutPositionOffsetX / Screen.width;
+                    _spoutScreenOffset.y = PluginConfig.Instance.AvatarSpoutPositionOffsetY / Screen.height;
+                    Graphics.Blit(_spoutTexture, dest, _spoutMaterial);
                 }
             }
         }